Relocate function code in own namespaced package
