+++ src/main/java/io/mathdojo/useraccountservice/model/interfaces/AccountHolder.java
@@ -1,0 +1,5 @@
+package io.mathdojo.useraccountservice.model.interfaces;
+
+public interface AccountHolder {
+
+}
