+++ src/main/java/io/mathdojo/useraccountservice/security/HTTPRequestSignatureVerifier.java
@@ -1,4 +1,4 @@
-package io.mathdojo.security;
+package io.mathdojo.useraccountservice.security;
 
 import java.io.UnsupportedEncodingException;
 import java.security.InvalidKeyException;
